**Improved the telemetry of the Hazelcast.** The original commits can be found in the improved-telemetry branch, and this is a squash of those. Along with the minimal information of hazelcast version, java version, operating system, uptime, memory used, Number of clients, Number of clusters, etc. the Phonehome process now also collects the following information- - Count of hazelcast maps with backup read enabled, map store, eviction, at least one query cache, hot restart enabled, at least one index, WAN replication configured, and at least one custom attribute. - Count of various distributed data structures including sets, queues, multimaps, caches, list, ring buffer, topic, replicated maps, cardinality estimator, PN counter, Flake Id generator. - Throughput and Latency of MapLoaders and MapStores. - Details of the Cloud used for deployment of hazelcast (AWS, Azure, or GCP), also whether it runs in Kubernetes or Docker.
Refactored the initial phonehome code with some necessary changes. Created a global node in the phonehome as the four methods(check() , phoneHome() , createParameters() and addClientInfo()) of the class received same node parameter.To avoid the complications in the PhoneHome class, a MetricsCollector Interface is used, which the various collectors implemented to collect relevant metrics. A common function of computeMetrics(Node node) to collect the metric details in Phonehome, made the design optimal to use.
Moving onto the integration tests for Phonehome using wire mock, it turned out that the network communication performed by PhoneHome can not be tested. Although if we fake the call home server with wire mock, the endpoint was hardcoded in the phonehome.So to achieve testability, the URL was made parametrizable. Also, to test the latency of maps with a map store, a fake map store(Delay Map Store) is being used, which gives a delay of 200ms to achieve testability. To achieve testability for the cloud details of Hazelcast, we made the CloudInfoCollector parametrized because we had to hardcode the endpoints of the cloud service providers.
PhoneHomeBenchmark measures the average time by the phonehome process for a node using quite a significant number of data structures(10k maps with different cases and 5k other distributed objects).
PhoneHomeMetrics enum contains all the details about the keys used for the various metrics for ease.
Following is the list of new metrics that are now being collected by Phonehome-
**Metrics name (Key)**
Number of Maps (mpct), Number of caches (cact), Number of caches with WAN replication configured (cawact), Number of maps with WAN replication configured (mpwact), Number of queues (quct), Number of multimaps (mmct), Number of sets (sect), Number of lists (lict), Number of ring buffers (rbct), Number of maps with backup read enabled (mpbrct), Number of maps using map store (mpmsct), Number of maps using at least one index (mpaoict), Number of maps having at least one custom attribute (mpaocct), Number of maps with at least one query cache (mpaoqcct), Numer of maps with hotstart enabled (mphect), Number of maps with eviction (mpevct), Number of maps with native in memory (mpnmct), Number of topics	(tpct), Number of replicated Map (rpct), Number of cardinality estimator (cect),Number of FlakeIdgenerator (figct), Number of PNcounter (pncct), Average Put Latency of map with mapstore (mpptlams), Average Put Latency of map without mapstore	(mpptla),Average Get Latency of map with mapstore (mpgtlams),Average Get Latency of map without mapstore (mpgtla), And the Cloud used for deployment of Hazelcast ( cld is G it means GCP, if A means AWS and if Z it means Azure) and whether uses docker( dck is K if in Kubernetes, if Only docker then it is D and is N if it is not in a docker).